docs: Merge get_dynamic_configuration
