+++ mirascope/core/base/_create.py
@@ -12,8 +12,7 @@ from ._utils import (
     get_possible_user_message_param,
 )
 from ._utils._get_dynamic_configuration import (
-    get_dynamic_configuration_async,
-    get_dynamic_configuration_sync,
+    get_dynamic_configuration,
 )
 from ._utils._protocols import fn_is_async
 from .call_params import BaseCallParams
@@ -94,7 +93,7 @@ def create_factory(  # noqa: ANN202
                 *args: _P.args, **kwargs: _P.kwargs
             ) -> TCallResponse | _ParsedOutputT:
                 fn_args = get_fn_args(fn, args, kwargs)
-                dynamic_config = await get_dynamic_configuration_async(fn, args, kwargs)
+                dynamic_config = await get_dynamic_configuration(fn, args, kwargs)
                 create, prompt_template, messages, tool_types, call_kwargs = setup_call(
                     model=model,
                     client=client,
@@ -134,7 +133,7 @@ def create_factory(  # noqa: ANN202
                 *args: _P.args, **kwargs: _P.kwargs
             ) -> TCallResponse | _ParsedOutputT:
                 fn_args = get_fn_args(fn, args, kwargs)
-                dynamic_config = get_dynamic_configuration_sync(fn, args, kwargs)
+                dynamic_config = get_dynamic_configuration(fn, args, kwargs)
                 create, prompt_template, messages, tool_types, call_kwargs = setup_call(
                     model=model,
                     client=client,
