+++ mirascope/core/base/_utils/_get_dynamic_configuration.py
@@ -1,5 +1,6 @@
+import inspect
 from collections.abc import Awaitable, Callable
-from typing import Any, TypeVar, cast
+from typing import Any, TypeVar, cast, overload
 
 from typing_extensions import TypeIs
 
@@ -15,40 +16,50 @@ def _is_message_param_list(result: object) -> TypeIs[list[BaseMessageParam]]:
     return isinstance(result, list)
 
 
-def get_dynamic_configuration_sync(
+@overload
+def get_dynamic_configuration(
+    fn: Callable[..., _BaseDynamicConfigT | list[BaseMessageParam]],
+    args: tuple[Any, ...],
+    kwargs: dict[str, Any],
+) -> _BaseDynamicConfigT: ...
+
+
+@overload
+def get_dynamic_configuration(
     fn: Callable[
-        ...,
-        _BaseDynamicConfigT | list[BaseMessageParam],
+        ..., Awaitable[_BaseDynamicConfigT] | Awaitable[list[BaseMessageParam]]
     ],
     args: tuple[Any, ...],
     kwargs: dict[str, Any],
-) -> _BaseDynamicConfigT:
-    result = cast(
-        _BaseDynamicConfigT | list[BaseMessageParam],
-        (
-            fn._original_fn(*args, **kwargs)  # pyright: ignore [reportFunctionMemberAccess]
-            if hasattr(fn, "_original_fn")
-            else fn(*args, **kwargs)
-        ),
-    )
-    if _is_message_param_list(result):
-        return cast(_BaseDynamicConfigT, {"messages": result})
-    return result
+) -> Awaitable[_BaseDynamicConfigT]: ...
 
 
-async def get_dynamic_configuration_async(
-    fn: Callable[..., Awaitable[_BaseDynamicConfigT | list[BaseMessageParam]]],
+def get_dynamic_configuration(
+    fn: Callable[
+        ...,
+        _BaseDynamicConfigT
+        | list[BaseMessageParam]
+        | Awaitable[_BaseDynamicConfigT]
+        | Awaitable[list[BaseMessageParam]],
+    ],
     args: tuple[Any, ...],
     kwargs: dict[str, Any],
-) -> _BaseDynamicConfigT:
-    result = cast(
-        _BaseDynamicConfigT | list[BaseMessageParam],
-        await (
-            fn._original_fn(*args, **kwargs)  # pyright: ignore [reportFunctionMemberAccess]
-            if hasattr(fn, "_original_fn")
-            else fn(*args, **kwargs)
-        ),
+) -> _BaseDynamicConfigT | Awaitable[_BaseDynamicConfigT]:
+    result = (
+        fn._original_fn(*args, **kwargs)  # pyright: ignore [reportFunctionMemberAccess]
+        if hasattr(fn, "_original_fn")
+        else fn(*args, **kwargs)
     )
+    if inspect.isawaitable(result):
+
+        async def inner_async() -> _BaseDynamicConfigT:
+            async_result = await result
+            if _is_message_param_list(async_result):
+                return cast(_BaseDynamicConfigT, {"messages": async_result})
+            return async_result
+
+        return inner_async()
+
     if _is_message_param_list(result):
         return cast(_BaseDynamicConfigT, {"messages": result})
     return result
