+++ tests/core/base/test_prompt.py
@@ -1,17 +1,14 @@
 """Tests for the `base_prompt` module."""
 
 import os
-from typing import Any, ClassVar, cast
-from unittest.mock import AsyncMock, MagicMock, Mock
+from typing import ClassVar
+from unittest import mock
+from unittest.mock import AsyncMock, MagicMock
 
 import pytest
-from PIL import Image
 from pydantic import computed_field
 
 from mirascope.core import BaseMessageParam
-from mirascope.core.base import ImagePart, TextPart
-from mirascope.core.base.dynamic_config import DynamicConfigMessages
-from mirascope.core.base.messages import Messages
 from mirascope.core.base.prompt import BasePrompt, metadata, prompt_template
 
 
@@ -235,157 +232,13 @@ def test_metadata_decorator() -> None:
     assert hasattr(fn, "_metadata") and fn._metadata == {"tags": {"version:0001"}}  # pyright: ignore [reportFunctionMemberAccess]
 
 
-def test_prompt_template_str_return():
-    @prompt_template()
-    def recommend_book(genre: str) -> str:
-        return f"recommend a {genre} book"
+def test_prompt_template_with_none() -> None:
+    """Tests the `prompt_template` decorator with `None` arguments."""
+    with mock.patch("mirascope.core.base.prompt._messages_decorator") as mock_decorator:
+        mock_decorated_function = mock.MagicMock()
+        mock_decorator.return_value.return_value = mock_decorated_function
 
-    result = recommend_book("fantasy")
-    assert isinstance(result, list)
-    assert len(result) == 1
-    assert isinstance(result[0], BaseMessageParam)
-    assert result[0].role == "user"
-    assert result[0].content == "recommend a fantasy book"
+        @prompt_template()
+        def fn() -> None: ...
 
-
-def test_prompt_template_dynamic_configuration_return():
-    @prompt_template()
-    def recommend_book(genre: str) -> DynamicConfigMessages:
-        return cast(DynamicConfigMessages, {"messages": f"recommend a {genre} book"})
-
-    result = recommend_book("fantasy")
-    assert isinstance(result, dict)
-    assert "messages" in result
-    assert result["messages"] == "recommend a fantasy book"
-
-
-def test_prompt_template_with_none_argument_str_return():
-    @prompt_template()
-    def recommend_book(genre: str) -> str:
-        return f"recommend a {genre} book"
-
-    result = recommend_book("fantasy")
-    assert isinstance(result, list)
-    assert len(result) == 1
-    assert isinstance(result[0], BaseMessageParam)
-    assert result[0].role == "user"
-    assert result[0].content == "recommend a fantasy book"
-
-
-@pytest.mark.asyncio
-async def test_prompt_template_str_return_async():
-    @prompt_template()
-    async def recommend_book(genre: str) -> str:
-        return f"recommend a {genre} book"
-
-    result = await recommend_book("fantasy")
-    assert isinstance(result, list)
-    assert len(result) == 1
-    assert isinstance(result[0], BaseMessageParam)
-    assert result[0].role == "user"
-    assert result[0].content == "recommend a fantasy book"
-
-
-@pytest.mark.asyncio
-async def test_prompt_template_dynamic_configuration_return_async():
-    @prompt_template()
-    async def recommend_book(genre: str) -> DynamicConfigMessages:
-        return cast(DynamicConfigMessages, {"messages": f"recommend a {genre} book"})
-
-    result = cast(DynamicConfigMessages, await recommend_book("fantasy"))
-    assert isinstance(result, dict)
-    assert "messages" in result
-    assert result["messages"] == "recommend a fantasy book"
-
-
-def test_list_str_return():
-    @prompt_template()
-    def recommend_book(genre: str) -> list[str]:
-        return ["hello!", f"recommend a {genre} book"]
-
-    result = recommend_book("fantasy")
-    assert isinstance(result, list)
-    assert len(result) == 1
-    assert isinstance(result[0], BaseMessageParam)
-    assert result == [
-        BaseMessageParam(
-            role="user",
-            content=[
-                TextPart(type="text", text="hello!"),
-                TextPart(type="text", text="recommend a fantasy book"),
-            ],
-        )
-    ]
-
-
-def test_list_message_role_return():
-    @prompt_template()
-    def recommend_book(genre: str) -> Messages.Type:
-        return [
-            Messages.System("You are a librarian"),
-            Messages.Assistant("I can help you find a book"),
-            Messages.User(
-                content=[
-                    TextPart(text="recommend a", type="text"),
-                    TextPart(text=genre, type="text"),
-                    TextPart(text="book", type="text"),
-                ]
-            ),
-        ]
-
-    result = recommend_book("fantasy")
-    assert isinstance(result, list)
-    assert len(result) == 3
-    assert isinstance(result[0], BaseMessageParam)
-    assert result[0].role == "system"
-    assert result[0].content == "You are a librarian"
-    assert result[1].role == "assistant"
-    assert result[1].content == "I can help you find a book"
-    assert isinstance(result[2], BaseMessageParam)
-    assert result[2].role == "user"
-    assert result[2].content == [
-        TextPart(type="text", text="recommend a"),
-        TextPart(type="text", text="fantasy"),
-        TextPart(type="text", text="book"),
-    ]
-
-
-def test_base_message_param_return():
-    @prompt_template()
-    def recommend_book(genre: str) -> BaseMessageParam:
-        return BaseMessageParam(role="user", content=f"hello! recommend a {genre} book")
-
-    result = recommend_book("fantasy")
-    assert isinstance(result, list)
-    assert len(result) == 1
-    assert isinstance(result[0], BaseMessageParam)
-    assert result[0].role == "user"
-    assert result[0].content == "hello! recommend a fantasy book"
-
-
-@pytest.fixture
-def mock_image():
-    return Mock(spec=Image.Image)
-
-
-def test_multimodal_return(mock_image):
-    mock_image.tobytes.return_value = b"\xff\xd8\xff"  # JPEG magic number
-
-    @prompt_template()
-    def recommend_book(previous_book: Image.Image) -> list[Any]:
-        return ["I just read this book:", previous_book, "What should I read next?"]
-
-    result = recommend_book(mock_image)
-
-    assert isinstance(result, list)
-    assert len(result) == 1
-    assert isinstance(result[0], BaseMessageParam)
-    assert result[0].role == "user"
-    assert len(result[0].content) == 3
-    assert result[0].content[0] == TextPart(type="text", text="I just read this book:")
-    assert result[0].content[1] == ImagePart(
-        type="image", media_type="image/jpeg", image=b"\xff\xd8\xff", detail=None
-    )
-    assert result[0].content[2] == TextPart(
-        type="text", text="What should I read next?"
-    )
+        assert fn == mock_decorated_function
